requirements.txt
# Sygnif — Freqtrade strategy dependencies
# Core (provided by freqtrade docker image)
# freqtrade
# pandas
# numpy
# ta-lib

# Sygnif terminal CLI (`sygnif_cli.py` / `sygnif` command)
rich>=13.9.0

# Additional strategy dependencies
feedparser>=6.0.11
pandas_ta>=0.3.14
requests>=2.32.0

# Bybit official Python SDK (scripts, research — not baked into Freqtrade image unless Dockerfile installs it)
pybit>=5.15.0

# Test dependencies (dev extras in requirements-dev.txt)
pytest>=8.3.0

# Lint / coverage / parallel tests (optional): pip install -r requirements-dev.txt

# Optional: Cursor MCP (mcp_rethink/server.py)
# mcp>=1.2.0
